# openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectedServicesRequest": {
        "description": "Request to resolve live connected services for an org/user principal.",
        "example": {
          "org_id": "org_12345",
          "user_id": "user_67890"
        },
        "properties": {
          "org_id": {
            "description": "Organization ID",
            "title": "Org Id",
            "type": "string"
          },
          "user_id": {
            "description": "User ID",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "user_id"
        ],
        "title": "ConnectedServicesRequest",
        "type": "object"
      },
      "ConnectedServicesResponse": {
        "description": "Lightweight connector-truth response for live connected/expired services.",
        "example": {
          "connected_services": [
            "quickbooks",
            "stripe"
          ],
          "expired_services": [
            "xero"
          ]
        },
        "properties": {
          "connected_services": {
            "description": "Services with currently usable credentials",
            "items": {
              "type": "string"
            },
            "title": "Connected Services",
            "type": "array"
          },
          "expired_services": {
            "description": "Services present but currently expired",
            "items": {
              "type": "string"
            },
            "title": "Expired Services",
            "type": "array"
          }
        },
        "title": "ConnectedServicesResponse",
        "type": "object"
      },
...
      "ConnectorStatusRequest": {
        "description": "Request to check connector status for specific services",
        "example": {
          "org_id": "org_12345",
          "services": [
            "quickbooks",
            "gmail",
            "vision"
          ],
          "user_id": "user_67890"
        },
        "properties": {
          "org_id": {
            "description": "Organization ID",
            "title": "Org Id",
            "type": "string"
          },
          "services": {
            "description": "List of services to check (e.g., ['quickbooks', 'gmail'])",
            "items": {
              "type": "string"
            },
            "title": "Services",
            "type": "array"
          },
          "user_id": {
            "description": "User ID",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "user_id",
          "services"
        ],
        "title": "ConnectorStatusRequest",
        "type": "object"
      },
      "ConnectorStatusResponse": {
        "description": "Response with connector status for specific services",
        "example": {
          "all_connected": false,
          "connectors": [
            {
              "credential_type": "customer",
              "display_name": "QuickBooks",
              "kind": "quickbooks",
              "last_updated": "2025-10-30T10:00:00",
              "requires_oauth": true,
              "status": "connected",
              "token_expiry": "2025-12-31T23:59:59"
            },
            {
              "credential_type": "agent",
              "display_name": "Gmail",
              "kind": "gmail",
              "requires_oauth": true,
              "status": "missing"
            }
          ],
          "missing_count": 1
        },
        "properties": {
          "all_connected": {
            "description": "True if all required connectors are authenticated",
            "title": "All Connected",
            "type": "boolean"
          },
          "connectors": {
            "description": "Status of each requested connector",
            "items": {
              "$ref": "#/components/schemas/WorkflowConnectorStatus"
            },
            "title": "Connectors",
            "type": "array"
          },
          "missing_count": {
            "description": "Number of connectors that need authentication",
            "title": "Missing Count",
            "type": "integer"
          }
        },
        "required": [
          "connectors",
          "all_connected",
          "missing_count"
        ],
        "title": "ConnectorStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowConnectorStatus": {
        "description": "Status for a single connector in workflow context",
        "example": {
          "credential_type": "customer",
          "display_name": "QuickBooks",
          "kind": "quickbooks",
          "last_updated": "2025-10-30T10:00:00",
          "requires_oauth": true,
          "status": "connected",
          "token_expiry": "2025-12-31T23:59:59"
        },
        "properties": {
          "credential_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "'agent' or 'customer' or None",
            "title": "Credential Type"
          },
          "display_name": {
            "description": "Human-readable name (e.g., 'QuickBooks')",
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "description": "Service name (e.g., 'quickbooks', 'gmail')",
            "title": "Kind",
            "type": "string"
          },
          "last_updated": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When credential was last updated",
            "title": "Last Updated"
          },
          "requires_oauth": {
            "description": "Whether this service requires OAuth",
            "title": "Requires Oauth",
            "type": "boolean"
          },
          "status": {
            "description": "Auth status: 'connected', 'expired', 'missing'",
            "title": "Status",
            "type": "string"
          },
          "token_expiry": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When token expires (if applicable)",
            "title": "Token Expiry"
          }
        },
        "required": [
          "kind",
          "display_name",
          "status",
          "requires_oauth"
        ],
        "title": "WorkflowConnectorStatus",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Execution Layer ('The Hands') for Aleq MIND",
    "title": "Stargate Lite",
    "version": "0.11.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/connectors/connected": {
      "post": {
        "description": "Return the lightweight live connector-truth surface for an org/user principal.",
        "operationId": "get_connected_services_api_v1_connectors_connected_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-Signature-256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Signature-256"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectedServicesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectedServicesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Connected Services",
        "tags": [
          "connectors"
        ]
      }
    },
    "/api/v1/connectors/status": {
      "post": {
        "description": "Check connector authentication status for specific services (workflow context).\n\nThis endpoint is called by the workflow manifest builder to check which\nconnectors are authenticated before presenting a workflow preview to the user.",
        "operationId": "check_workflow_connector_status_api_v1_connectors_status_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-Signature-256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Signature-256"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Workflow Connector Status",
        "tags": [
          "connectors"
        ]
      }
    },
...
  }
}